Finish shrinking export files to one per exporter
