+++ errai-common-apt/src/main/java/org/jboss/errai/common/apt/generator/ExportedTypesFromSource.java
@@ -47,4 +47,8 @@ public class ExportedTypesFromSource {
   public Set<Element> findAnnotatedSourceElements(final TypeElement typeElement) {
     return new HashSet<>(exportedTypes.get(typeElement.getQualifiedName().toString()));
   }
+
+  public boolean isEmpty() {
+    return exportedTypes.isEmpty();
+  }
 }
