+++ errai-common-apt/src/test/java/org/jboss/errai/apt/internal/export/any_other_Module__test__ExportFile.java
@@ -16,11 +16,15 @@
 
 package org.jboss.errai.apt.internal.export;
 
-import org.jboss.errai.common.apt.exportfile.ExportedTypesFromExportFilesTestExportedType;
+import org.jboss.errai.common.apt.exportfile.TestExportedType;
 
 /**
  * @author Tiago Bento <tfernand@redhat.com>
  */
-public class any_Module__test__ExportFile_org_jboss_errai_common_apt_exportfile_ExportedTypesFromExportFilesTestAnnotation {
-  ExportedTypesFromExportFilesTestExportedType sa7H9hdg;
+public class any_other_Module__test__ExportFile {
+
+  public class org_jboss_errai_common_apt_exportfile_TestAnnotation {
+
+    TestExportedType randomName;
+  }
 }
