+++ errai-common-apt/src/test/java/org/jboss/errai/apt/internal/export/org_jboss_errai_common_apt_localapps_localapp1_module1_TestModule1__test__ExportFile.java
@@ -16,11 +16,13 @@
 
 package org.jboss.errai.apt.internal.export;
 
-import org.jboss.errai.common.apt.exportfile.TestExportedType;
+import org.jboss.errai.common.apt.localapps.localapp1.module1.TestExportedType1;
 
 /**
  * @author Tiago Bento <tfernand@redhat.com>
  */
-public class any_other_Module__test__ExportFile_org_jboss_errai_common_apt_exportfile_TestAnnotation {
-  TestExportedType randomName;
+public class org_jboss_errai_common_apt_localapps_localapp1_module1_TestModule1__test__ExportFile {
+  public class org_jboss_errai_common_apt_exportfile_ExportedTypesFromExportFilesTestAnnotation {
+    TestExportedType1 testExportedType1;
+  }
 }
