+++ errai-common-apt/src/test/java/org/jboss/errai/apt/internal/export/org_jboss_errai_common_apt_localapps_localapp2_module2_TestModule2__test__ExportFile.java
@@ -16,12 +16,15 @@
 
 package org.jboss.errai.apt.internal.export;
 
-import org.jboss.errai.common.apt.exportfile.TestExportedType;
-import org.jboss.errai.common.apt.localapps.localapp1.module1.TestExportedType1;
+import org.jboss.errai.common.apt.localapps.localapp2.module2.TestExportedType21;
+import org.jboss.errai.common.apt.localapps.localapp2.module2.TestExportedType22;
 
 /**
  * @author Tiago Bento <tfernand@redhat.com>
  */
-public class org_jboss_errai_common_apt_localapps_localapp1_module1_TestModule1__test__ExportFile_org_jboss_errai_common_apt_exportfile_ExportedTypesFromExportFilesTestAnnotation {
-  TestExportedType1 testExportedType1;
+public class org_jboss_errai_common_apt_localapps_localapp2_module2_TestModule2__test__ExportFile {
+  public class org_jboss_errai_common_apt_exportfile_ExportedTypesFromExportFilesTestAnnotation {
+    TestExportedType21 testExportedType21;
+    TestExportedType22 testExportedType22;
+  }
 }
