+++ errai-common-apt/src/test/java/org/jboss/errai/common/apt/exportfile/ErraiAptPackagesTest.java
@@ -16,9 +16,8 @@
 
 package org.jboss.errai.common.apt.exportfile;
 
-import org.jboss.errai.apt.internal.export.any_other_Module__test__ExportFile_org_jboss_errai_common_apt_exportfile_TestAnnotation;
+import org.jboss.errai.apt.internal.export.any_other_Module__test__ExportFile;
 import org.jboss.errai.codegen.apt.test.ErraiAptTest;
-import org.jboss.errai.common.apt.exportfile.ErraiAptPackages;
 import org.junit.Test;
 
 import static org.junit.Assert.assertTrue;
@@ -36,9 +35,7 @@ public class ErraiAptPackagesTest extends ErraiAptTest {
   @Test
   public void testElementsInPackage() {
     assertTrue(ErraiAptPackages.exportFilesPackageElement(elements)
-            .map(p -> p.getEnclosedElements()
-                    .contains(getTypeElement(
-                            any_other_Module__test__ExportFile_org_jboss_errai_common_apt_exportfile_TestAnnotation.class)))
+            .map(p -> p.getEnclosedElements().contains(getTypeElement(any_other_Module__test__ExportFile.class)))
             .orElse(false));
   }
 }
