--- errai-common-apt/src/test/java/org/jboss/errai/common/apt/exportfile/ExportFileNameTest.java
@@ -1,53 +1,0 @@
-/*
- * Copyright (C) 2017 Red Hat, Inc. and/or its affiliates.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *       http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.jboss.errai.common.apt.exportfile;
-
-import org.jboss.errai.codegen.apt.test.ErraiAptTest;
-import org.junit.Test;
-
-import java.lang.annotation.Annotation;
-
-import static java.util.Collections.emptySet;
-import static org.jboss.errai.common.apt.exportfile.ExportFileName.decodeAnnotationClassNameFromExportFileName;
-import static org.jboss.errai.common.apt.exportfile.ExportFileName.encodeAnnotationNameAsExportFileName;
-import static org.junit.Assert.assertEquals;
-
-/**
- * @author Tiago Bento <tfernand@redhat.com>
- */
-public class ExportFileNameTest extends ErraiAptTest {
-
-  @interface TestAnnotationInner {
-  }
-
-  @Test
-  public void testEncodeDecode() {
-    testEncodeDecode(TestAnnotation.class);
-  }
-
-  @Test
-  public void testEncodeDecodeInner() {
-    testEncodeDecode(TestAnnotationInner.class);
-  }
-
-  private void testEncodeDecode(final Class<? extends Annotation> testAnnotationClass) {
-    final ExportFile1 exportFile = new ExportFile1("test", getTypeElement(testAnnotationClass), emptySet());
-    final String encodedName = encodeAnnotationNameAsExportFileName(exportFile);
-    assertEquals(testAnnotationClass.getCanonicalName(), decodeAnnotationClassNameFromExportFileName(encodedName));
-  }
-
-}
